# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Life Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Weather {
  /** Temperature in Celsius */
  temperature: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  description: string;
  /** Weather condition code (e.g. "clear", "cloudy", "rainy") */
  icon: string;
  city: string;
}

export interface NewsArticle {
  id: string;
  title: string;
  source: string;
  url: string;
  publishedAt: string;
  /** @nullable */
  description?: string | null;
}

export interface Todo {
  id: number;
  text: string;
  completed: boolean;
  createdAt: string;
}

export interface TodoInput {
  /** @minLength 1 */
  text: string;
}

export interface TodoUpdate {
  /** @minLength 1 */
  text?: string;
  completed?: boolean;
}

export interface Quote {
  text: string;
  author: string;
}

export type GetWeatherParams = {
  /**
   * Latitude (defaults to New York if omitted)
   */
  lat?: number;
  /**
   * Longitude
   */
  lon?: number;
};