# lib/api-zod/src/generated/types/todoUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Life Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TodoUpdate {
  /** @minLength 1 */
  text?: string;
  completed?: boolean;
}