# lib/api-zod/src/generated/types/weather.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Life Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Weather {
  /** Temperature in Celsius */
  temperature: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  description: string;
  /** Weather condition code (e.g. "clear", "cloudy", "rainy") */
  icon: string;
  city: string;
}